Changing kill script to be more robust, adding logger

+++ src/package/execserver/bin/azkaban-executor-shutdown.sh
@@ -6,6 +6,14 @@ echo "Shutting down current running AzkabanExecutorServer at port $executorport"
 
 proc=`cat $azkaban_dir/currentpid`
 
-kill $proc
+if ! [ "$proc" ]; then
+    ps -ef |grep "java.*azkaban-exec-server" | grep -v grep
+    proc=`ps -ef |grep "java.*azkaban-exec-server" | grep -v grep | sed 's/  */ /g' | cut -d' ' -f2`
+fi
+
+
+if [ "$proc" ]; then
+    kill $proc
+fi
 
 cat /dev/null > $azkaban_dir/currentpid
